receiver/pyproject.toml
[tool.poetry]
name = "receiver"
version = "0.1.0"
description = ""
authors = ["kaiosatiro <kaiosatiro@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.4.0"
tenacity = "^9.0.0"
fastapi = "^0.112.0"
httpx = "^0.27.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
fastapi-cli = "^0.0.5"
pyjwt = "^2.9.0"
pika = "^1.3.2"
asgi-correlation-id = "^4.3.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
